requirements.txt
gspread
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
moviepy
requests
Pillow
python-dotenv

